+++ packages/agent-core/src/Agent/Tools/Ghci/Runtime.hs
@@ -14,6 +14,7 @@ module Agent.Tools.Ghci.Runtime
     ) where
 
 import Agent.Cancel (CancelFlag, isCancelled, waitCancel)
+import Agent.OsPath (unsafeToFilePath)
 import Agent.Tools.Ghci.Classify
     ( GhciClass(..)
     , classifyGhciInput
@@ -50,7 +51,6 @@ import Control.Concurrent.STM
 import Control.Exception.Safe
     ( SomeException
     , finally
-    , impureThrow
     , mask
     , onException
     , try
@@ -85,7 +85,6 @@ import System.Process
     )
 import System.Posix.Signals (sigINT, signalProcessGroup)
 import System.Posix.Types (ProcessGroupID)
-import System.OsPath (OsPath, decodeUtf)
 
 data GhciOutcome
     = GhciCompleted
@@ -468,7 +467,7 @@ ghciArgs =
 spawnProcess :: ToolEnv -> IO (Either Text GhciProcess)
 spawnProcess env = do
     let spec = (proc "ghci" ghciArgs)
-            { cwd = Just (decodeUtfPath env.toolCwd)
+            { cwd = Just (unsafeToFilePath env.toolCwd)
             , std_in = CreatePipe
             , std_out = CreatePipe
             , std_err = CreatePipe
@@ -869,6 +868,3 @@ remainingMillis started budget = do
     now <- getMonotonicTimeNSec
     let elapsed = fromIntegral ((now - started) `div` 1000000)
     pure (max 0 (budget - elapsed))
-
-decodeUtfPath :: OsPath -> FilePath
-decodeUtfPath = either impureThrow id . decodeUtf